requirements.txt
Django==1.9.6

dj-database-url==0.4.1
django-allauth==0.25.2
django-autoslug==1.9.3
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-extensions==1.6.7
-e git+https://github.com/Tivix/django-rest-auth.git@3189a5c7a0be6540f8b57642cce71176879205e2#egg=django_rest_auth
django-rest-swagger==0.3.7
django-storages-redux==1.3.2
django-taggit-serializer==0.1.5
django-taggit==0.18.3
djangorestframework==3.3.3
drf-extensions==0.2.8

python-decouple==3.0

boto==2.40.0
Collectfast==0.2.3
Pillow==3.2.0
psycopg2==2.6.1
waitress==0.9.0

coverage==4.0.3
flake8==2.5.4
model-mommy==1.2.6
